update search, steer, gemmascope tests
