+++ apps/webapp/tests-playwright/fixtures.ts
@@ -3,6 +3,8 @@ import { test as base } from '@playwright/test';
 type MyFixtures = {
     baseUrl: string;
     gemmaUrl: string;
+    steerUrl: string;
+    searchUrl: string;
 };
 
 export const test = base.extend<MyFixtures>({
@@ -10,4 +12,10 @@ export const test = base.extend<MyFixtures>({
   gemmaUrl: async ({ baseUrl }, use) => {
     await use(`${baseUrl}/gemma-scope`);
   },
+  steerUrl: async ({ baseUrl }, use) => {
+    await use(`${baseUrl}/steer`)
+  },
+  searchUrl: async ({ baseUrl }, use) => {
+    await use(`${baseUrl}/search`);
+  }
 });
